+++ ChangeLog.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ### Changed
 - Numerical input from PDG 2018
 - Running of strong coupling (consistent decoupling at flavor thresholds)
+- example.py includes an example for Wilson coefficients specified in electroweak unbroken theory
 
 ### Fixed
 - Bug in running of the strong coupling constant (decoupling at wrong loop order)